MenuEditCol Componenet
