+++ app/components/Sheet/ColumnOptions.jsx
@@ -3,6 +3,7 @@ import classNames from 'classnames/bind';
 import styles from 'css/components/table';
 import { DropdownButton, Glyphicon, Dropdown } from 'react-bootstrap';
 import { MenuItem } from 'react-bootstrap';
+import MenuEditCol from './MenuEditCol';
 
 const cx = classNames.bind(styles);
 
@@ -18,33 +19,22 @@ export default class ColumnOptions extends Component {
 		this.sortAsc = this.sortAsc.bind(this);
 		this.sortDec = this.sortDec.bind(this);
 		this.exitTypeMenu = this.exitTypeMenu.bind(this);
-		this.saveTypeChanges = this.saveTypeChanges.bind(this);
-		this.itemSelected = this.itemSelected.bind(this);
 	}
 
 	handleSelection(evt, evtKey){
 		this[evtKey]();
 	}
 
-	itemSelected(e, ekey) {
-		console.log('itemSelected', ekey);
-		this.setState({selectedType: ekey});
-	}
-
 	exitTypeMenu() {
-		console.log('exitTypeMenu');
+		console.log('ColOpt exitTypeMenu');
 		this.setState({view: 'dropdown'});
-	}
-
-	saveTypeChanges() {
-		console.log('saveTypeChanges');
-		this.exitTypeMenu();
 	}
 
 	changeType() {
 		console.log('changeType');
 		this.setState({view: 'editNameAndType'});
 	}
+
 
 	rename() {
 		console.log('rename');
@@ -91,35 +81,7 @@ export default class ColumnOptions extends Component {
 			    </Dropdown>)
 
 		} else if (this.state.view === 'editNameAndType') {
-
-			let columnTypes = ['Text','Number','Checkbox','Select'];
-
-			function generateTypes () {
-				return columnTypes.map((type, idx) =>
-				{
-					return (<MenuItem key={idx+1} eventKey={type}>{type}</MenuItem>);
-				})
-			}
-
-			viewing = (
-				<div className={cx('editNameAndType')}>
-					<div className={cx('thead') + ' col-md-12'} contentEditable>{this.props.data.name}</div>
-					<Dropdown id="dropdown-custom-1" onSelect={this.itemSelected} className={cx('typeDropdown') + ' col-md-12'}>
-				      <Dropdown.Toggle noCaret className=' col-md-12'>
-				        {this.state.selectedType || this.props.data.type} <Glyphicon className={cx('columnCarrat')} glyph="menu-down" />
-				      </Dropdown.Toggle>
-				      <Dropdown.Menu className={cx('columnMenu')}>
-				      	{generateTypes()}
-				      </Dropdown.Menu>
-				    </Dropdown>
-
-				    <p className='col-md-12'> A single line of text. You can optionally prefill each cell with a default value: </p>
-				    <div className='col-md-12'>
-					    <button className="btn col-md-5" type="button" onClick={this.exitTypeMenu}>Cancel</button>
-					    <button className="btn btn-primary col-md-5" type="button" onClick={this.saveTypeChanges}>Save</button>
-					</div>
-				</div>
-				)
+			viewing = (<MenuEditCol data={this.props.data} exitTypeMenu={this.exitTypeMenu}/>)
 		}
 
 		return (
