+++ app/css/components/table.css
@@ -16,7 +16,6 @@
 
 .editNameAndType {
   z-index: 99;
-  margin-top: 10px;
   padding: 10px;
   width: 250px;
   position: absolute;
